refactor: adjustments of code review
