+++ app/src/main/java/com/codandotv/streamplayerapp/navigation/NavigationGraph.kt
@@ -16,7 +16,7 @@ import com.codandotv.streamplayerapp.core_navigation.routes.BottomNavRoutes
 import com.codandotv.streamplayerapp.core_navigation.routes.Routes
 import com.codandotv.streamplayerapp.feature_list_streams.detail.presentation.navigation.detailStreamNavGraph
 import com.codandotv.streamplayerapp.feature_list_streams.list.presentation.navigation.listStreamsNavGraph
-import com.codandotv.streamplayerapp.feature_list_streams.profile.presentation.navigation.profilePickerStreamNavGraph
+import com.codandotv.streamplayerapp.feature_profile.profile.presentation.navigation.profilePickerStreamNavGraph
 import com.codandotv.streamplayerapp.splah.presentation.navigation.splashNavGraph
 
 @Composable
