+++ core-shared/src/main/java/com/codandotv/streamplayerapp/core_shared/extension/String.Ext.kt
@@ -1,0 +1,3 @@
+package com.codandotv.streamplayerapp.core_shared.extension
+
+fun String.Companion.empty() = ""
