--- feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/profile/data/model/ProfileStreamResponse.kt
@@ -1,11 +1,0 @@
-package com.codandotv.streamplayerapp.feature_list_streams.profile.data.model
-
-data class ProfileStreamResponse(
-    val id: String,
-    val name: String,
-    val profile_url: String,
-)
-
-data class ProfilesResponse(
-    val profiles: List<ProfileStreamResponse>
-)
