--- feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/profile/domain/ProfileStream.kt
@@ -1,7 +1,0 @@
-package com.codandotv.streamplayerapp.feature_list_streams.profile.domain
-
-data class ProfileStream(
-    val id: String = "",
-    val name: String = "",
-    val imageUrl: String = "",
-)
