--- feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/profile/presentation/navigation/ProfilePickerStreamNavigation.kt
@@ -1,29 +1,0 @@
-package com.codandotv.streamplayerapp.feature_list_streams.profile.presentation.navigation
-
-import androidx.lifecycle.Lifecycle
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.composable
-import com.codandotv.streamplayerapp.core_navigation.routes.BottomNavRoutes.PARAM.PROFILE_ID
-import com.codandotv.streamplayerapp.core_navigation.routes.BottomNavRoutes.HOME
-import com.codandotv.streamplayerapp.core_navigation.routes.Routes
-import com.codandotv.streamplayerapp.feature_list_streams.profile.di.ProfilePickerStreamModule
-import com.codandotv.streamplayerapp.feature_list_streams.profile.presentation.screens.ProfilePickerStreamScreen
-import org.koin.core.context.loadKoinModules
-import org.koin.core.context.unloadKoinModules
-
-fun NavGraphBuilder.profilePickerStreamNavGraph(navController: NavHostController) {
-    composable(Routes.PROFILE_PICKER) { nav ->
-        if (nav.getLifecycle().currentState == Lifecycle.State.STARTED) {
-            loadKoinModules(ProfilePickerStreamModule.module)
-        }
-        ProfilePickerStreamScreen(
-            navController = navController,
-            onNavigateListStreams = { profilePic ->
-                navController.navigate("$HOME?$PROFILE_ID=$profilePic")
-            }
-        ) {
-            unloadKoinModules(ProfilePickerStreamModule.module)
-        }
-    }
-}
