--- feature-list-streams/src/main/java/com/codandotv/streamplayerapp/feature_list_streams/profile/presentation/widget/ProfilePickerStreamLoad.kt
@@ -1,20 +1,0 @@
-package com.codandotv.streamplayerapp.feature_list_streams.profile.presentation.widget
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-
-
-@Composable
-fun LoadScreen() {
-    Box(Modifier.fillMaxSize()) {
-        CircularProgressIndicator(
-            modifier = Modifier.align(
-                Alignment.Center
-            )
-        )
-    }
-}
