+++ feature-profile/src/main/java/com/codandotv/streamplayerapp/feature_profile/profile/domain/ProfilePickerStreamMapper.kt
@@ -1,0 +1,11 @@
+package com.codandotv.streamplayerapp.feature_profile.profile.domain
+
+import com.codandotv.streamplayerapp.feature_profile.profile.data.model.ProfilesResponse
+
+fun ProfilesResponse.toProfiles(): List<ProfileStream> = this.profiles.map { profileResponse ->
+    ProfileStream(
+        id = profileResponse.id,
+        name = profileResponse.name,
+        imageUrl = profileResponse.profile_url,
+    )
+}
