+++ feature-profile/src/main/java/com/codandotv/streamplayerapp/feature_profile/profile/domain/ProfileStream.kt
@@ -1,0 +1,7 @@
+package com.codandotv.streamplayerapp.feature_profile.profile.domain
+
+data class ProfileStream(
+    val id: String,
+    val name: String,
+    val imageUrl: String
+)
