+++ feature-profile/src/main/res/drawable/netflix_horizontal_logo.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="200.dp"
+    android:height="54dp"
+    android:viewportWidth="1024"
+    android:viewportHeight="278">
+  <path
+      android:pathData="M140.8,258.9c-15.4,2.7 -31.1,3.5 -47.3,5.7l-49.5,-144.9v151.1c-15.4,1.6 -29.5,3.8 -44.1,5.9v-276.7h41.1l56.2,157v-157h43.5v258.9zM225.9,101.3c16.8,0 42.4,-0.8 57.8,-0.8v43.2c-19.2,0 -41.6,0 -57.8,0.8v64.3c25.4,-1.6 50.8,-3.8 76.5,-4.6v41.6l-119.7,9.5v-255.4h119.7v43.2h-76.5v58.1zM463.2,43.2h-44.9v198.9c-14.6,0 -29.2,0 -43.2,0.5v-199.4h-44.9v-43.2h133l-0,43.2zM533.5,98.4h59.2v43.2h-59.2v98.1h-42.4v-239.7h120.8v43.2h-78.4v55.1zM682.1,201.9c24.6,0.5 49.5,2.4 73.5,3.8v42.7c-38.6,-2.4 -77.3,-4.9 -116.8,-5.7v-242.7h43.2v201.9zM792.1,251.3c13.8,0.8 28.4,1.6 42.4,3.2v-254.6h-42.4v251.3zM1024,0l-54.9,131.6 54.9,145.1c-16.2,-2.2 -32.4,-5.1 -48.6,-7.8l-31.1,-80 -31.6,73.5c-15.7,-2.7 -30.8,-3.5 -46.5,-5.7l55.7,-126.8 -50.3,-130h46.5l28.4,72.7 30.3,-72.7h47.3z"
+      android:fillColor="#E50813"/>
+</vector>
