+++ feature-profile/src/main/res/values/strings.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="profile_animation_background_opacity">Fundo da tela escurecendo</string>
+    <string name="profile_animation_selected_image_size">Imagem de perfil selecionada aumentando de tamanho</string>
+    <string name="profile_animation_selected_image_position"> "Imagem de perfil selecionada se movendo para o centro"</string>
+    <string name="profile_animation_showing_all_profiles">Mostrando todos os perfis</string>
+    <string name="profile_current_profile_name">Imagem de perfil de %s</string>
+</resources>
